Fix build37 get_variants for range queries
Unfortunately, `get_variants` currently mutates the input `query`
parameter and sets in it the `genomicBuild` without unsetting it.

Since we no longer normalise variants to have them both in build37
and build38, then the final query will end up looking for build38
while all the provided variants are in build37.