Fix bug in get_variant and fix tests
Need to run the Flask server in a background thread so the
seqfetcher endpoint is available for the hgvs library to call.
Details here: https://stackoverflow.com/questions/77676757/run-a-full-flask-server-under-pytest-to-handle-http-requests-on-a-given-port

Also need to pick the relevat transcript in `process_NC_HGVS()` in
such a way that we select the one with the most up to date version
so we don't end up unable to find the RefSeq data because it looks
like we don't have all the RNA transcripts that UTA knows about.